TEC-6257 move the exception to the exception package
