+++ src/main/java/com/echobox/api/linkedin/exception/LinkedInJsonMappingException.java
@@ -17,8 +17,6 @@
 
 package com.echobox.api.linkedin.exception;
 
-import com.echobox.api.linkedin.jsonmapper.LinkedInException;
-
 /**
  * LinkedIn Json mapping exception
  * @author Joanna
